api_test/requirements.txt
openai==0.27.8
python-dotenv==0.19.1
